+++ src/main/java/bc1/gream/domain/order/service/helper/coupon/CouponCalculator.java
@@ -1,4 +1,4 @@
-package bc1.gream.domain.order.service.helper;
+package bc1.gream.domain.order.service.helper.coupon;
 
 import bc1.gream.domain.user.entity.Coupon;
 import bc1.gream.global.common.ResultCase;
